src/administration/admins/api/serializers.py
from rest_framework import serializers

from src.administration.admins.models import Product


class ProductSerializer(serializers.ModelSerializer):

    class Meta:
        model = Product
        fields = '__all__'

