requirements.txt
six~=1.16.0
Faker
pillow
markdown
python-dateutil

django~=3.2.16
django-environ
django-filter~=21.1
django-allauth
django-resized
python-decouple
django-ckeditor
django-crispy-forms


django-browser-reload

djangorestframework~=3.12.4
DateTime~=5.0